+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - [github.com/InVisionApp/go-logger](https://github.com/InVisionApp/go-logger) integration
+- `simplelogadapter` to make logger library integration easier
 
 ### Changed